requirements.txt
flask==3.0.3
flask-cors==4.0.1
werkzeug==3.0.3
langdetect==1.0.9
gtts==2.5.1
pyttsx3==2.90
PyPDF2==3.0.1
python-docx==1.2.0
python-pptx==0.6.23
pytesseract==0.3.13
Pillow==10.4.0
requests==2.32.3
pydub==0.25.1
playsound==1.3.0